Rename camoRequirements to camouflageRequirements
