+++ src/data/camouflageRequirements.js
